Address PR review feedback for OpenRouter realm sync
- Fix .json suffix bug in fetchExistingSlugs that caused slug mismatches
- Remove incorrect lastSeenInApi update when marking models as deprecated
- Fix error message to match field name (realmUrl not realmURL)
- Fix format helpers treating 0 as missing (use == null instead of !num)
- Register openrouter_realm matrix user in register-realm-users.sh
- Extract cron jobs from worker-manager.ts into lib/cron-scheduler.ts
